+++ structurizr-core/src/com/structurizr/model/DefaultImpliedRelationshipsStrategy.java
@@ -1,0 +1,13 @@
+package com.structurizr.model;
+
+/**
+ * The default strategy is to NOT create implied relationships.
+ */
+public class DefaultImpliedRelationshipsStrategy extends AbstractImpliedRelationshipsStrategy {
+
+    @Override
+    public void createImpliedRelationships(Relationship relationship) {
+        // do nothing
+    }
+
+}